Actualización datos Entidades.

+++ API/Models/BaseProveedorDTO.cs
@@ -2,5 +2,5 @@ public class BaseProveedorDTO
 {
     public string Nombre { get; set; }
     public string Poblacion { get; set; }
-    public decimal Telefono { get; set; }
+    public int Telefono { get; set; }
 }
